Simplify some logics in ReplicaCntBalancer and move RangeLeaderBalancer.
