+++ bifromq-inbox/bifromq-inbox-store/src/main/resources/META-INF/services/com.baidu.bifromq.basekv.balance.IStoreBalancerFactory
@@ -11,7 +11,6 @@
 # See the License for the specific language governing permissions and limitations under the License.
 #
 
-com.baidu.bifromq.inbox.store.balance.RecoveryBalancerFactory
 com.baidu.bifromq.inbox.store.balance.ReplicaCntBalancerFactory
 com.baidu.bifromq.inbox.store.balance.RangeSplitBalancerFactory
 com.baidu.bifromq.inbox.store.balance.RangeLeaderBalancerFactory
